DeepSleepRatio Double -> Int로 수정

+++ app/src/main/java/com/snoffee/app/domain/model/SleepData.kt
@@ -7,6 +7,6 @@ data class SleepData(
     val date: Long,                         // 수면 날짜 (PK)
     val sleepStart: Long,                   // 수면 시작 시각
     val sleepEnd: Long,                     // 수면 종료 시각
-    val deepSleepRatio: Double,              // 깊은 수면 비율 (0.0 ~ 1.0)
+    val deepSleepRatio: Int,              // 깊은 수면 비율 (0.0 ~ 1.0)
     val source: String = "manual"
 )
